+++ CHANGELOG.md
@@ -35,6 +35,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## v0.17.4
+
+- Cache no candidate or ranked routes
+
+## v0.17.3
+
+- Skip pool filtering if min osmo liquidity is zero
+
 ## v0.17.2
 
 - Fix bug with max split routes parameter